requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langsmith==0.3.42
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.38.2
numpy==2.2.5
openai==1.77.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
protobuf==6.30.2
pyarrow==20.0.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pydotenv==0.0.7
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
storage3==0.11.3
streamlit==1.45.0
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websockets==14.2
yarl==1.20.0
zstandard==0.23.0
